semantic_layer/requirements.txt
numpy
pandas
google-generativeai
fastapi
fastmcp
uvicorn
starlette
weaviate-client
pydantic
